* commit 'b4cf2a2aee78b0ed3e75b5a07baaf5479fe708d9':
  Convert Java/JNI to 64-bit, part 2.

+++ graphics/java/android/renderscript/AllocationAdapter.java
@@ -26,12 +26,12 @@ import android.util.TypedValue;
  *
  **/
 public class AllocationAdapter extends Allocation {
-    AllocationAdapter(int id, RenderScript rs, Allocation alloc) {
+    AllocationAdapter(long id, RenderScript rs, Allocation alloc) {
         super(id, rs, alloc.mType, alloc.mUsage);
         mAdaptedAllocation = alloc;
     }
 
-    int getID(RenderScript rs) {
+    long getID(RenderScript rs) {
         throw new RSInvalidStateException(
             "This operation is not supported with adapters at this time.");
     }
